requirements.txt
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
live2d-py>=0.3.2
numpy>=1.21.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
